+++ other/buildfiles/rpgenginepreparer/src/Main.java
@@ -7,7 +7,7 @@ import java.util.Collections;
 public class Main {
 
     private static FTPTools ftp = null;
-    private final static String launcherVersion = "8", playVersion = "1.11.6", createVersion = "1.11.5";
+    private final static String launcherVersion = "8", playVersion = "1.11.6", createVersion = "1.11.6";
 
     public static void main(String[] args) {
 
@@ -17,10 +17,10 @@ public class Main {
 
         connectToFTP();
 
-        createAndUploadDocumentation();
-        //createAndUploadLauncher();
         //createAndUploadCreateOrPlay("create", createVersion);
         createAndUploadCreateOrPlay("play", playVersion);
+        createAndUploadDocumentation();
+        //createAndUploadLauncher();
 
         cleanUp();